requirements.txt
torch==2.7.1+cu128
torchvision==0.22.1+cu128
opencv-python==4.11.0.86
pillow==11.3.0
timm==1.0.22
einops==0.8.1
numpy==1.25.0
scipy==1.15.3
pandas==1.3.5
scikit-learn==1.7.2
matplotlib==3.10.6
tensorboard==2.20.0
grad-cam==1.5.5
tqdm==4.65.2
thop==0.1.1.post2209072238
torchstat==0.0.7
torchinfo==1.8.0
torchsummary==1.5.1